Setup the basics for asynchronous messaging
* Added an artemis docker image to the docker-compose.yml
* Added a consumer to the appliance service to consume workflow messages
* Added a producer to the home service to produce workflow messages

Note the services are not actually doing anything yet with those consumers/producers, but was just trying to get the plumbing setup and connected to the artemis "machine". That being said their are some unused variables and what not that will get resolved in future MRs.

Also, sonar (and intellij) are complaining about the session and consumer being in a try-with-resources.  Unfortunately, when I added them to the same try it errored because of the order of operations. (e.g. clientID needs to be the first action on connection, before you can call createSession and createConnection does not accept the ID in the constructor).  I really didn't want to have nested try blocks, but at the same time these probably are not being closed correctly.