Added `@NotNull` annotations to the register calls per review
I only added `@NotNull` instead of `@Valid` because I haven't added any validations to the objects themselves. If I get around to it, maybe I will, but it is not critical at this point since these are not real services.

Also fixed an issue where the state, color, and brightness were not getting the default values. This is due to the lombok annotations and I don't care too much about fixing that since this is only used for a test runner.